Update, bug fixes, feature request.
Updated to 1.4.7.
Fixed a bug where an item would be duplicated (infact one with stack
size 1 in chest A and the other with stack size 0 in chest B) when
colliding on the same tick with two chests.
Introduced a configuration file that allows to modify the pull distance
of the vortex device.